Added/mofied honeypots and network sensors, modified logstash confs

+++ scripts/install_smtp.sh
@@ -2,11 +2,19 @@
 
 set -x
 set -e
+
+read -p "Setup E-mail alerts [Y/N]" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+
 yum remove sendmail -y
 yum install postfix -y
 
-sed -i "s/#mydomain = domain.tld/mydomain = $domainName /g" /etc/postfix/main.cf
-sed -i 's/#myorigin = $mydomain/myorigin = $mydomain/g' /etc/postfix/main.cf
+echo "virtual_alias_domains = $domainName
+virtual_alias_maps = hash:/etc/postfix/virtual" >> /etc/postfix/main.cf
+
+ead -p "Enter e-mail to forward all alerts to: " -e mail
+echo "@$domainName $email" >> /etc/postfix/virtual
 
 systemctl enable postfix
 systemctl start postfix
+fi
